Contact.js was renamed Footer.js
